+++ project/src/components/place-map/place-map.jsx
@@ -1,25 +1,10 @@
-import React, {useEffect, useRef} from 'react';
-import PropTypes from 'prop-types';
+import React, {useRef} from 'react';
 import useMap from '../../hooks/useMap/useMap';
-import leaflet from 'leaflet';
-import {MAP_ICON_DEFAULT} from '../../const';
+import PropTypes from 'prop-types';
 
 function PlaceMap({points}) {
   const mapRef = useRef(null);
-  const city = points[0].city;
-  const map = useMap(mapRef, city);
-
-  useEffect(() => {
-    if (map) {
-      points.forEach((point) => {
-        const {location: {latitude, longitude}} = point;
-        const marker = leaflet.marker([latitude, longitude], {
-          icon: leaflet.icon(MAP_ICON_DEFAULT),
-        });
-        marker.addTo(map);
-      });
-    }
-  }, [map, points]);
+  useMap(mapRef, points, null);
 
   return (
     <section className="property__map map">
@@ -31,6 +16,7 @@ function PlaceMap({points}) {
 PlaceMap.propTypes = {
   points: PropTypes.arrayOf(
     PropTypes.shape({
+      id: PropTypes.number.isRequired,
       city: PropTypes.shape({
         location: PropTypes.shape({
           latitude: PropTypes.number.isRequired,
