+++ project/src/hooks/useMap/useMap.js
@@ -1,8 +1,12 @@
 import {useEffect, useState} from 'react';
 import leaflet from 'leaflet';
+import 'leaflet/dist/leaflet.css';
+import {MAP_ICON_DEFAULT, MAP_ICON_ACTIVE} from '../../const';
 
-function useMap(mapRef, city) {
+function useMap(mapRef, points, activePoint) {
   const [map, setMap] = useState(null);
+  const city = points[0].city;
+  const markers = [];
 
   useEffect(() => {
     if (mapRef !== null) {
@@ -30,6 +34,23 @@ function useMap(mapRef, city) {
       setMap(instance);
     }
   }, [mapRef, map, city]);
+
+  useEffect(() => {
+    if (map) {
+      points.forEach((point) => {
+        const {id, location: {latitude, longitude}} = point;
+        const marker = leaflet.marker([latitude, longitude], {
+          icon: (activePoint !== null && activePoint.id === id)
+            ? leaflet.icon(MAP_ICON_ACTIVE)
+            : leaflet.icon(MAP_ICON_DEFAULT),
+        });
+        marker.addTo(map);
+        markers.push(marker);
+      });
+
+      return () => markers.forEach((marker) => marker.remove());
+    }
+  }, [map, points, activePoint]);
 
   return map;
 }
